feat: Send captured image to server
Added the functionality to send captured images to the server. The `sendToServer` function converts a base64 image into a Blob object and sends it to the server through a POST request. After sending, it displays a success message and clears the captured image. In case of error, displays an error message.